+++ samples/dialogs/dialogs/dialogs.vcxproj
@@ -60,7 +60,7 @@ del /s /q /f $(IntDir)\*.oelf</NMakeCleanCommandLine>
   <ItemDefinitionGroup>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="main.cpp" />
+    <ClCompile Include="main.c" />
     <ClCompile Include="build.bat" />
   </ItemGroup>
   <ItemGroup>
